マップの配列は共通処理として切り出す
