+++ packages/tiled-map/src/TiledMap.ts
@@ -1,125 +1,7 @@
-import { MapChip, MapChipProperties,  AutoTileMapChipProperties, isAutoTileMapChipProperties, AutoTileMapChip } from './MapChip'
-import { MapChipImage, MapChipsCollection, MapChipCollectionProperties, MapChipImageProperties } from './MapChips'
+import { MapChip } from './MapChip'
+import { MapChipsCollection, MapChipCollectionProperties } from './MapChips'
 import { AutoTiles, AutoTilesProperties } from './AutoTile/AutoTiles'
-
-export type TiledMapDataProperties = {
-  chipCountX: number,
-  chipCountY: number,
-  mapData: Array<MapChipProperties | AutoTileMapChipProperties | null>
-}
-
-export class TiledMapData {
-  constructor(
-    private _chipCountX: number,
-    private _chipCountY: number,
-    private _mapData: Array<MapChip | null> = []
-  ) {
-    if (this._mapData.length > 0 && this.size !== this._mapData.length) {
-      throw new Error()
-    }
-
-    if (this._mapData.length === 0) this.allocate()
-  }
-
-  get size() {
-    return this._chipCountX * this._chipCountY
-  }
-
-  get width() {
-    return this._chipCountX
-  }
-
-  get height() {
-    return this._chipCountY
-  }
-
-  get mapData() {
-    return this._mapData
-  }
-
-  set(mapData: Array<MapChip | null>) {
-    if (mapData.length !== this._mapData.length) throw new Error()
-
-    this._mapData = mapData
-  }
-
-  transferFromTiledMapData(src: TiledMapData, srcX: number, srcY: number, width: number, height: number, destX: number, destY: number) {
-    for(let x = 0; x < width; x++) {
-      const putX = destX + x
-      const pickupX = srcX + x
-      if (putX < 0 || putX >= this.width) continue;
-      if (pickupX < 0 || pickupX >= src.width) continue;
-
-      for(let y = 0; y < height; y++) {
-        const putY = destY + y
-        const pickupY = srcY + y
-        if (putY < 0 || putY >= this.height) continue;
-        if (pickupY < 0 || pickupY >= src.height) continue;
-
-        this.put(src.getMapDataFromChipPosition(pickupX, pickupY), putX, putY)
-      }
-    }
-  }
-
-  filter(needles: Array<MapChip>): TiledMapData {
-    const filtered = this._mapData.map(chip => needles.some(needle => !!chip && needle.compare(chip)) ? chip : null)
-    return new TiledMapData(
-      this.width,
-      this.height,
-      filtered
-    )
-  }
-
-  getMapDataFromChipPosition(x: number, y: number): MapChip | null {
-    if ((x < 0)  || (y < 0) || (x >= this._chipCountX) || (y >= this._chipCountY)) return null
-
-    const mapNumber = this.convertPositionToMapNumber(x, y)
-    return this._mapData[mapNumber]
-  }
-
-  put(mapChip: MapChip | null, x: number, y: number) {
-    const mapNumber = this.convertPositionToMapNumber(x, y)
-    this._mapData[mapNumber] = mapChip
-  }
-
-  convertPositionToMapNumber(x: number, y: number) {
-    return y * this._chipCountX + x
-  }
-
-  convertMapNumberToPosition(num: number) {
-    return {
-      x: num % this._chipCountY,
-      y: Math.floor(num / this._chipCountY)
-    }
-  }
-
-  private allocate() {
-    this._mapData = new Array(this._chipCountY * this._chipCountX).fill(null)
-  }
-
-  toObject(): TiledMapDataProperties {
-    return {
-      chipCountX: this._chipCountX,
-      chipCountY: this._chipCountY,
-      mapData: this._mapData.map(data => data?.toObject() || null)
-    }
-  }
-
-  static fromObject(val: TiledMapDataProperties) {
-    const mapData = val.mapData.map(data => {
-      if (!data) return null
-
-      if (isAutoTileMapChipProperties(data)) {
-        return AutoTileMapChip.fromObject(data)
-      }
-
-      return MapChip.fromObject(data)
-
-    })
-
-    return new TiledMapData(val.chipCountX, val.chipCountY, mapData)
-  }
-}
+import { TiledMapData, TiledMapDataProperties } from './MapData/TiledMapData'
 
 export type TiledMapProperties = {
   chipCountX: number,
